+++ src/media_centers/kodi/TvShowXmlReader.cpp
@@ -34,12 +34,45 @@ void TvShowXmlReader::parseNfoDom(QDomDocument domDoc)
     if (!domDoc.elementsByTagName("showtitle").isEmpty()) {
         m_show.setShowTitle(domDoc.elementsByTagName("showtitle").at(0).toElement().text());
     }
-    if (!domDoc.elementsByTagName("rating").isEmpty()) {
-        m_show.setRating(domDoc.elementsByTagName("rating").at(0).toElement().text().replace(",", ".").toDouble());
-    }
-    if (!domDoc.elementsByTagName("votes").isEmpty()) {
-        m_show.setVotes(
-            domDoc.elementsByTagName("votes").at(0).toElement().text().replace(",", "").replace(".", "").toInt());
+    // check for new ratings syntax
+    if (!domDoc.elementsByTagName("ratings").isEmpty()) {
+        // <ratings>
+        //   <rating name="default" default="true">
+        //     <value>10</value>
+        //     <votes>10</votes>
+        //   </rating>
+        // </ratings>
+        auto ratings = domDoc.elementsByTagName("ratings").at(0).toElement().elementsByTagName("rating");
+        for (int i = 0; i < ratings.length(); ++i) {
+            Rating rating;
+            auto ratingElement = ratings.at(i).toElement();
+            rating.source = ratingElement.attribute("name", "default");
+            rating.rating =
+                ratingElement.elementsByTagName("value").at(0).toElement().text().replace(",", ".").toDouble();
+            rating.voteCount = ratingElement.elementsByTagName("votes")
+                                   .at(0)
+                                   .toElement()
+                                   .text()
+                                   .replace(",", "")
+                                   .replace(".", "")
+                                   .toInt();
+            m_show.ratings().push_back(rating);
+            m_show.setChanged(true);
+        }
+
+    } else if (!domDoc.elementsByTagName("rating").isEmpty()) {
+        // otherwise use "old" syntax:
+        // <rating>10.0</rating>
+        // <votes>10.0</votes>
+        Rating rating;
+        rating.rating = domDoc.elementsByTagName("rating").at(0).toElement().text().replace(",", ".").toDouble();
+        if (!domDoc.elementsByTagName("votes").isEmpty()) {
+            rating.voteCount =
+                domDoc.elementsByTagName("votes").at(0).toElement().text().replace(",", "").replace(".", "").toInt();
+        }
+        m_show.ratings().clear();
+        m_show.ratings().push_back(rating);
+        m_show.setChanged(true);
     }
     if (!domDoc.elementsByTagName("top250").isEmpty()) {
         m_show.setTop250(domDoc.elementsByTagName("top250").at(0).toElement().text().toInt());